Add exception handling. Fixes #4
Validations shouldn't throw exceptions because of
type mismatches and the like. If the rule is 'Length(5)' and
the value in the field is 5, that should be a validation failure,
not a TypeError because you can't call len() on an int.
It's not ideal to have to hide exceptions like this because
there could be actual problems with a validator, but we're just going
to have to rely on tests preventing broken things.